refactor : 공통된 `addressToString` 이동
