+++ src/main/java/com/example/tripKo/domain/place/entity/Place.java
@@ -2,6 +2,7 @@ package com.example.tripKo.domain.place.entity;
 
 import com.example.tripKo.domain.BaseTimeEntity;
 import com.example.tripKo.domain.address.entity.Address;
+import com.example.tripKo.domain.address.entity.AddressCategory;
 import com.example.tripKo.domain.contents.entity.Contents;
 import com.example.tripKo.domain.contents.entity.ContentsMenu;
 import com.example.tripKo.domain.file.entity.File;
@@ -67,6 +68,13 @@ public class Place extends BaseTimeEntity {
 
     public void addContents(Contents contents) {
         this.contents.add(contents);
+    }
+
+    public String addressToString(Address address) {
+        String addressToString = address.getBuildingName() + " " + address.getRoadName();
+        AddressCategory addressCategory = address.getAddressCategory();
+        String addressCategoryToString = addressCategory.getEmdName() + " " + addressCategory.getSiggName() + " " + addressCategory.getSidoName();
+        return addressToString + " " + addressCategoryToString;
     }
 
 }
